requirements.txt
numpy==1.26.4
Pillow==10.4.0
h5py==3.12.1
tensorflow==2.16.2
opencv-python==4.10.0.84
configparser